+++ R/piping.R
@@ -144,7 +144,19 @@
       .cur <- try(eval(.bracketExpression, envir=envir), silent=TRUE)
       if (inherits(.cur, "try-error")) {
       } else if (length(.cur) > 1) {
-        if (identical(.cur[[1]], quote(`{`))) {
+        if (inherits(.cur, "character")) {
+          if (is.null(names(.cur))) {
+            .cur <- lapply(.cur, function(x) {
+              str2lang(x)
+            })
+          } else {
+            .cur <- lapply(names(.cur), function(x) {
+              str2lang(paste0(x, "<-", .cur[[x]]))
+            })
+          }
+          .cur <- as.call(c(list(quote(`{`)),.cur))
+          .bracketExpression <- .cur
+        } else if (identical(.cur[[1]], quote(`{`))) {
           .bracketExpression <- .cur
         }
       }